Add files via upload
This is a clone of the 2048 game. The basic idea for the implementation of this game was to study the algorithm of the game to get a better idea. These files are study notes of what I learned throughout the process of understanding the algorithm of the 2048 game and it is not a work of my own.